+++ backend/src/main/java/br/com/inthurn/backend/model/transport/response/CashBalanceResponseDTO.java
@@ -1,4 +1,6 @@
 package br.com.inthurn.backend.model.transport.response;
 
-public record CashBalanceResponseDTO (String id, String description, Double initialValue){
+import java.math.BigDecimal;
+
+public record CashBalanceResponseDTO (String id, String description, BigDecimal initialValue){
 }
